# docs/swagger.yaml
info:
  contact: {}
paths:
  /:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 首页
      tags:
      - 登录模块
  /attach/upload:
    post:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 上传图片
  /contact/addfriend:
    post:
      parameters:
      - description: userId
        in: formData
        name: userId
        type: string
      - description: targetId
        in: formData
        name: targetId
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加好友
      tags:
      - 个人中心
  /contact/createCommunity:
    post:
      parameters:
      - description: ownerId
        in: query
        name: ownerId
        type: string
      - description: name
        in: query
        name: name
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 创建群聊
      tags:
      - 用户模块
  /contact/loadCommunity:
    post:
      parameters:
      - description: ownerId
        in: query
        name: ownerId
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 加载群聊
      tags:
      - 用户模块
  /searchFriends:
    post:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 查找好友
      tags:
      - 用户模块
  /toChat:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 个人界面
      tags:
      - 登录模块
  /toRegister:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 注册界面
      tags:
      - 登录模块
  /user/creatUser:
    get:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/deleteUser:
    get:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/findUserByNameAndPwd:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 登录
      tags:
      - 用户模块
  /user/getUserList:
    get:
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户列表
      tags:
      - 用户模块
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改个人信息
      tags:
      - 用户模块
swagger: "2.0"